requirements.txt
streamlit
python-dotenv
keybert
transformers
pandas
scikit-learn
nltk
dateparser
python-dateutil
seaborn